# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/admin/login": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Login",
        "description": "Resolve admin credentials to a session.\n\nOnly providers that hold credentials themselves (``static``) support this;\nOAuth and token-verifier providers return 401. On success the bearer token\nis stored in an HttpOnly session cookie so the browser UI never persists\nit; the token stays in the response for non-browser API clients.",
        "operationId": "login_api_admin_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Login Api Admin Login Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/session": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Establish Session",
        "description": "Adopt a browser-flow or pasted bearer token into the session cookie.\n\nThe browser Authorization Code + PKCE flow and the pasted-token flow\nterminate in JavaScript, so the UI hands the resulting token back to the\ngateway, which verifies it and stores it in an HttpOnly cookie.",
        "operationId": "establish_session_api_admin_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Establish Session Api Admin Session Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "End Session",
        "description": "Clear the admin session cookie.",
        "operationId": "end_session_api_admin_session_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        }
      }
    },
...
    "/api/admin/servers": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Servers",
        "description": "Return the currently mounted servers.",
        "operationId": "list_servers_api_admin_servers_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ServerView"
                  },
                  "type": "array",
                  "title": "Response List Servers Api Admin Servers Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Create Server",
        "description": "Create and live-mount a server.",
        "operationId": "create_server_api_admin_servers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServerCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/servers/{name}": {
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Update Server",
        "description": "Replace a server's definition live; the name is immutable.",
        "operationId": "update_server_api_admin_servers__name__put",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServerUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Delete Server",
        "description": "Unmount and delete a server.",
        "operationId": "delete_server_api_admin_servers__name__delete",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Auth0AuthProviderConfig": {
        "properties": {
          "resource_base_url": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resource Base Url"
          },
          "issuer_url": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Issuer Url"
          },
          "allowed_client_redirect_uris": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowed Client Redirect Uris"
          },
          "jwt_signing_key": {
            "anyOf": [
              {
                "type": "string",
                "format": "password",
                "writeOnly": true
              },
              {
                "type": "null"
              }
            ],
            "title": "Jwt Signing Key"
          },
          "require_authorization_consent": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "string",
                "enum": [
                  "remember",
                  "external"
                ]
              }
            ],
            "title": "Require Authorization Consent",
            "default": true
          },
          "consent_csp_policy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Consent Csp Policy"
          },
          "forward_resource": {
            "type": "boolean",
            "title": "Forward Resource",
            "default": true
          },
          "fallback_refresh_token_expiry_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Refresh Token Expiry Seconds"
          },
          "fastmcp_access_token_expiry_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Fastmcp Access Token Expiry Seconds"
          },
          "token_expiry_threshold_seconds": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Token Expiry Threshold Seconds",
            "default": 0
          },
          "required_scopes": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Required Scopes"
          },
          "provider": {
            "type": "string",
            "const": "auth0",
            "title": "Provider"
          },
          "config_url": {
            "type": "string",
            "minLength": 1,
            "format": "uri",
            "title": "Config Url"
          },
          "client_id": {
            "type": "string",
            "title": "Client Id"
          },
          "client_secret": {
            "type": "string",
            "format": "password",
            "title": "Client Secret",
            "writeOnly": true
          },
          "audience": {
            "type": "string",
            "title": "Audience"
          },
          "timeout_seconds": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout Seconds",
            "default": 10
          },
          "redirect_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Redirect Path"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "provider",
          "config_url",
          "client_id",
          "client_secret",
          "audience"
        ],
        "title": "Auth0AuthProviderConfig"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KeycloakAuthProviderConfig": {
        "properties": {
          "provider": {
            "type": "string",
            "const": "keycloak",
            "title": "Provider"
          },
          "realm_url": {
            "type": "string",
            "minLength": 1,
            "format": "uri",
            "title": "Realm Url"
          },
          "required_scopes": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Required Scopes"
          },
          "audience": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audience"
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "provider",
          "realm_url"
        ],
        "title": "KeycloakAuthProviderConfig"
      },
      "LoginRequest": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest"
      },
...
      "ServerCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "pattern": "^[a-z0-9][a-z0-9-]*$",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "maxLength": 255,
            "title": "Description",
            "default": ""
          },
          "upstream_url": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 1,
            "format": "uri",
            "title": "Upstream Url"
          },
          "auth": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Auth0AuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/AuthKitAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/AwsCognitoAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/AzureAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/DescopeAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/DiscordAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/GitHubAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/GoogleAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/HuggingFaceAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/JwtAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/KeycloakAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/OciAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/PropelAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/ScalekitAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/SupabaseAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/WorkOsAuthProviderConfig"
              }
            ],
            "title": "Auth",
            "discriminator": {
              "propertyName": "provider",
              "mapping": {
                "auth0": "#/components/schemas/Auth0AuthProviderConfig",
                "authkit": "#/components/schemas/AuthKitAuthProviderConfig",
                "aws-cognito": "#/components/schemas/AwsCognitoAuthProviderConfig",
                "azure": "#/components/schemas/AzureAuthProviderConfig",
                "descope": "#/components/schemas/DescopeAuthProviderConfig",
                "discord": "#/components/schemas/DiscordAuthProviderConfig",
                "github": "#/components/schemas/GitHubAuthProviderConfig",
                "google": "#/components/schemas/GoogleAuthProviderConfig",
                "huggingface": "#/components/schemas/HuggingFaceAuthProviderConfig",
                "jwt": "#/components/schemas/JwtAuthProviderConfig",
                "keycloak": "#/components/schemas/KeycloakAuthProviderConfig",
                "oci": "#/components/schemas/OciAuthProviderConfig",
                "propelauth": "#/components/schemas/PropelAuthProviderConfig",
                "scalekit": "#/components/schemas/ScalekitAuthProviderConfig",
                "supabase": "#/components/schemas/SupabaseAuthProviderConfig",
                "workos": "#/components/schemas/WorkOsAuthProviderConfig"
              }
            }
          },
          "verify_upstream_tls": {
            "type": "boolean",
            "title": "Verify Upstream Tls",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "upstream_url",
          "auth"
        ],
        "title": "ServerCreateRequest",
        "description": "Payload for creating a server."
      },
...
      "ServerUpdateRequest": {
        "properties": {
          "description": {
            "type": "string",
            "maxLength": 255,
            "title": "Description",
            "default": ""
          },
          "upstream_url": {
            "type": "string",
            "maxLength": 2048,
            "minLength": 1,
            "format": "uri",
            "title": "Upstream Url"
          },
          "auth": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Auth0AuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/AuthKitAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/AwsCognitoAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/AzureAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/DescopeAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/DiscordAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/GitHubAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/GoogleAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/HuggingFaceAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/JwtAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/KeycloakAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/OciAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/PropelAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/ScalekitAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/SupabaseAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/WorkOsAuthProviderConfig"
              }
            ],
            "title": "Auth",
            "discriminator": {
              "propertyName": "provider",
              "mapping": {
                "auth0": "#/components/schemas/Auth0AuthProviderConfig",
                "authkit": "#/components/schemas/AuthKitAuthProviderConfig",
                "aws-cognito": "#/components/schemas/AwsCognitoAuthProviderConfig",
                "azure": "#/components/schemas/AzureAuthProviderConfig",
                "descope": "#/components/schemas/DescopeAuthProviderConfig",
                "discord": "#/components/schemas/DiscordAuthProviderConfig",
                "github": "#/components/schemas/GitHubAuthProviderConfig",
                "google": "#/components/schemas/GoogleAuthProviderConfig",
                "huggingface": "#/components/schemas/HuggingFaceAuthProviderConfig",
                "jwt": "#/components/schemas/JwtAuthProviderConfig",
                "keycloak": "#/components/schemas/KeycloakAuthProviderConfig",
                "oci": "#/components/schemas/OciAuthProviderConfig",
                "propelauth": "#/components/schemas/PropelAuthProviderConfig",
                "scalekit": "#/components/schemas/ScalekitAuthProviderConfig",
                "supabase": "#/components/schemas/SupabaseAuthProviderConfig",
                "workos": "#/components/schemas/WorkOsAuthProviderConfig"
              }
            }
          },
          "verify_upstream_tls": {
            "type": "boolean",
            "title": "Verify Upstream Tls",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "upstream_url",
          "auth"
        ],
        "title": "ServerUpdateRequest",
        "description": "Payload for updating a server; the name is immutable."
      },
      "ServerView": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "upstream_url": {
            "type": "string",
            "title": "Upstream Url"
          },
          "auth": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Auth0AuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/AuthKitAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/AwsCognitoAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/AzureAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/DescopeAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/DiscordAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/GitHubAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/GoogleAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/HuggingFaceAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/JwtAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/KeycloakAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/OciAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/PropelAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/ScalekitAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/SupabaseAuthProviderConfig"
              },
              {
                "$ref": "#/components/schemas/WorkOsAuthProviderConfig"
              }
            ],
            "title": "Auth",
            "discriminator": {
              "propertyName": "provider",
              "mapping": {
                "auth0": "#/components/schemas/Auth0AuthProviderConfig",
                "authkit": "#/components/schemas/AuthKitAuthProviderConfig",
                "aws-cognito": "#/components/schemas/AwsCognitoAuthProviderConfig",
                "azure": "#/components/schemas/AzureAuthProviderConfig",
                "descope": "#/components/schemas/DescopeAuthProviderConfig",
                "discord": "#/components/schemas/DiscordAuthProviderConfig",
                "github": "#/components/schemas/GitHubAuthProviderConfig",
                "google": "#/components/schemas/GoogleAuthProviderConfig",
                "huggingface": "#/components/schemas/HuggingFaceAuthProviderConfig",
                "jwt": "#/components/schemas/JwtAuthProviderConfig",
                "keycloak": "#/components/schemas/KeycloakAuthProviderConfig",
                "oci": "#/components/schemas/OciAuthProviderConfig",
                "propelauth": "#/components/schemas/PropelAuthProviderConfig",
                "scalekit": "#/components/schemas/ScalekitAuthProviderConfig",
                "supabase": "#/components/schemas/SupabaseAuthProviderConfig",
                "workos": "#/components/schemas/WorkOsAuthProviderConfig"
              }
            }
          },
          "verify_upstream_tls": {
            "type": "boolean",
            "title": "Verify Upstream Tls"
          }
        },
        "type": "object",
        "required": [
          "name",
          "description",
          "upstream_url",
          "auth",
          "verify_upstream_tls"
        ],
        "title": "ServerView",
        "description": "Admin-facing server representation with secrets masked."
      },
      "SessionRequest": {
        "properties": {
          "token": {
            "type": "string",
            "title": "Token"
          }
        },
        "type": "object",
        "required": [
          "token"
        ],
        "title": "SessionRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}